+++ NEWS.md
@@ -201,6 +201,10 @@ Notable changes in Lepton EDA 1.9.10 (upcoming)
   selection)`.  All the functions and variables available are still
   re-exported in the former one but its using is discouraged.
 
+- The module `(gschem symbol check)` has been renamed to `(schematic
+  symbol check)`.  All the functions and variables available are still
+  re-exported in the former one but its using is discouraged.
+
 - Grips can now be turned on and off at run-time.
   Use the new 'Options → Grips: On/Off' menu item or
   <kbd>O</kbd>-<kbd>I</kbd> keyboard shortcut.